test: separate pyarrow tests from other tests
